refactor(loop-host): split system inference tests
